cleaning up file structure and dockerfile
