+++ monitoring_service/examples/test_reading_sample_response_from_model_api.py
